requirements.txt
# Core framework
fastapi
uvicorn
pydantic

# LangGraph and LangChain
langgraph
langchain
langchain-community
langchain-core

# LiveKit real-time communication
livekit
livekit-api
livekit-agents
livekit-plugins-groq
livekit-plugins-silero
livekit-plugins-turn-detector
livekit-plugins-noise-cancellation
livekit-plugins-cartesia

# Google API clients
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
googleapis-common-protos

# HTTP and async clients
aiohttp
aiohttp-retry
httpx
requests

# LLM providers
groq
google-generativeai  # Gemini AI for email analysis

# Telephony
twilio

# Utilities
python-dotenv
python-dateutil
dateparser  # Natural language date/time parsing
pytz
typing-extensions
bs4
authlib  # For Google OAuth
itsdangerous  # For session management
jinja2  # For HTML templates

# Development and testing
pytest
pytest-asyncio
black
flake8

# Logging and monitoring
structlog
rich
